init create subscription workflow step
This step receives the IDs of the cart, order, and customer, along with
the subscription’s details.In this step, you use the createSubscriptions
method to create the subscription. In the metadata property, you set the
ID of the order created on purchase.The step returns the created
subscription as well as an array of links to create.